requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
jinja2==3.1.5
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
